Put player-specific application states in separate directories
