+++ src/redux/reducers/game.js
@@ -1,6 +1,6 @@
 import { types } from '../actions/game';
 import { Channel } from 'fmg-core';
-import * as playerA from '../../game-engine/application-states/ApplicationStatesPlayerA';
+import * as playerA from '../../game-engine/application-states/PlayerA/index';
 
 // Fake data for development purposes
 const gameLibrary = 0x111;
